SALTO-6213: Moving index.ts to package root
This saves the need for custom knip configuration for these packages and aligns better with our conventions in other packages.
