+++ packages/persistent-pool/package.json
@@ -11,10 +11,10 @@
     "access": "public"
   },
   "files": [
-    "dist/src"
+    "dist"
   ],
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "build": "../../build_utils/turbo_run.sh build-ts ; ../../build_utils/turbo_run.sh lint",
     "build-ts": "tsc -b",
